[vpnmodel] Ask connman to establish VPN connection. JB#42691
As much as possible (preferably everything) should go through connman.
This particular one solves the problem with connman not knowing when
the user cancels the connection (authentication dialog). The result
of the Connect call was being returned to the caller and the caller
was Settings, rather than connman who had no idea what was going on.