[systemsettings] Disable autoconnect before removing VPN. Contributes to JB#42797
This fixes the issue of removing a connected VPN from the list. It is
imperative to set the autoconnect flag false, then disconnect and after
that remove the VPN. This happens because ConnMan will try to reconnect
the VPN, fails in connecting and then saves the state of the VPN,
recreating configuration files of the deleted VPN if it is only
disconnected before removing it.