[nemo-storage] Make block devices without partition/partitiontable visible in model. Fixes JB#42943
This commit adds signleton block devices container that is the main
interface that monitor uses for operating block devices. The block devices
emits a signal (newBlock) when it sees an acceptable new block device and throws
away the ones that cannot / should not be shown in the model.

When block device container sees a block device that could be a container
for a partition, it'll wait for 3000ms milliseconds before accepting the
block. If partition appears during this inverval, then waiting is cancel.