[nemo-qml-plugin-systemsettings] Support VPN ordering by connected status. Contributes to JB#45380
This change adds the `orderByConnected` flag to VpnModel which, when
set, will order VPN connections in the model first by the connection
status (connected first, followed by not connected), then by the name
alphabetially ascending (case-sensitive).

When not set, the previous case-sensitive alphabetic name ordering is
used.

The new ordering is used when the MDM policy is set to prevent the user
from manually connecting or disconnecting the VPNs. In this case, the
list of VPNs is presented differently with headers indicating connection
status, rather than indicating it using a TextSwitch glassitem.