[systemsettings] Allow certificate model to decode in-memory PEMs. Contributes to JB#40089
The certificate model code allows an X509 certificate (or bundle of
certificates) to be loaded from disk so they can be used to the UI
(e.g. for displaying certificate details).

This change allows PEM data to be provided as a QByteArray, rather than
as a file, allowing certificates store in memory to be accessed by the
UI as well, without having to generate a temprory file.

This change also exposes certificate summary info, providing some
addition properties that help when displaying a summary of the
certificate.