Store locale.conf per user and set system locale.
[systemsettings] Store locale to path based on uid. Contributes to JB#47651
[systemsettings] Set system locale. Contributes to JB#47651

Store user and system locale in separate directories. Store user locale
based on uid instead of username since that is easier to handle in
certain service files. Setting device owner locale sets system locale
too.

Working with paths that may not exist is a little ugly in Qt and it
doesn't seem to have a way to set directory permissions at all.

Signed-off-by: Tomi Leppänen <tomi.leppanen@jolla.com>