[nemo-systemsettings] Add alone property to UserInfo. Contributes to JB#49696
Add alone property that can be used to determine if user is the only
user on device. It uses tristated value on the background so that if it
isn't needed, it's not checked. It is also wired up to the file system
watcher so that it gets updated if groups change.

Signed-off-by: Tomi Leppänen <tomi.leppanen@jolla.com>