Stop monitoring seat changes if uid is set
Technically it would be possible to create UserInfo object in QML before
activation has happened and then set it's uid to another user. In
practice that's quite unusual thing to do. Previously it would have
replaced the user with current user but now it just stops monitoring for
changes instead.

Signed-off-by: Tomi Leppänen <tomi.leppanen@jolla.com>