[partionmanager] Replace isExternal with hintAuto
Instead of staticly detecting if a device is external using a expose
Udisks2.Block.HintAuto attribute to do the same.

Udev set the HintAuto attribute on a device which should be
automounted. Since we only want to list in the storage ui devices which we are external,
which happen to be only those that one would auto mount we can use the
said attribute from Udisks.
In this case Udisks2 does all the work and no fancy thinking needed.
This also has the advantage that we simply can override this behavior by settings the udev attribute
UDISKS_AUTO on the block device in question (see man 8 udisksm)