[nemo-qml-plugin-systemsettings] Move location config to /var. Contributes to JB#56083
Sandboxing doesn't like to have files modified under /etc as private-etc
will copy the content on startup and later changes are not applied.

Moved the config under /var/lib/location/ instead, but the old path
is also getting written to keep compatibility with code that reads the
file directly, not via LocationSettings class.

This will reset the settings state. Would be nice to copy the old file,
but that should be done only once and this is also now packaging
the initial states so it's tricky.