[location] Prevent location state changes during emergency calls. JB#57710
AML is ran with the emergency calls and it requires to keep location
services, i.e., GPS on during the call. It must not be allowed for the
user to try to change this during the call.

This change tracks the call state with QMceCallState functionality. When
state is "Active" and type is "Emergency" all changes made to the
location or GPS status are prevented and warning is logged.