[location] Add canToggle properties for location and GPS. JB#57710
This adds the properties and signals for getting info whether the
location and GPS state can be modified. An enum is used to define the
level of access to enabling location&GPS with a bit more granularity.
Not all emergency calls are AML enabled calls that need explicit access
to location. As of now, we can just detect emergency call -> always
disabled when it is active.