[nemo-qml-plugin-systemsettings] Fix race on tracking Udisks2 jobs. JB#62852
The job status change tracking was done by Jobs themselves.
However, the signal tracking is done by the dbus-daemon, and
by the time it was started on that side, the job finished signal
might have had happened already and the result never got handled.

So instead make the Monitor class track all the jobs and then find
if there's an instance for it.

I suspect commit ce28996f8d might have introduced the
"Make sure job is completed." part as some sort of stop-gap for this
kind of failure. Not sure is that still needed and assuming success
is suspicious, but keeping it for now.