[nemo-qml-plugin-systemsettings] Make partition space check via statvfs64() async. JB#56182
Evidently at least after mounting statvfs64() might block for several
seconds. That's not a good thing to be executed via the UI.

So instead move the disk space checks to happen with a separate task.

Also switched partition APIs to return -1 for unknown values.
With 0 the app cannot know if it's proper value or something still to be
fetched.

Could still consider if the stat tasks should have only max one thread,
but the QThreadPool setup was nice and commonly avoids any unnecessary
threads.